+++ examples/auto_retry_order.py
@@ -2,7 +2,7 @@ import os
 from time import sleep
 
 import pyotp
-import robin_stocks as r
+import robin_stocks.robinhood as r
 from dotenv import load_dotenv
 
 '''
