+++ examples/two_factor_log_in.py
@@ -1,7 +1,7 @@
 import os
 
 import pyotp
-import robin_stocks as r
+import robin_stocks.robinhood as r
 from dotenv import load_dotenv
 '''
 This is an example script that will automatically log you in with two factor authentication.
