+++ examples/write_options_to_file.py
@@ -1,4 +1,4 @@
-import robin_stocks as r
+import robin_stocks.robinhood as r
 import os
 import datetime
 import time as t
