--- robin_stocks/__init__.py
@@ -1,158 +1,0 @@
-from .account import get_historical_portfolio,      \
-                     get_all_positions,             \
-                     get_bank_account_info,         \
-                     get_bank_transfers,            \
-                     get_card_transactions,         \
-                     unlink_bank_account,           \
-                     get_open_stock_positions,      \
-                     get_dividends,                 \
-                     get_total_dividends,           \
-                     get_dividends_by_instrument,   \
-                     get_notifications,             \
-                     get_latest_notification,       \
-                     get_linked_bank_accounts,      \
-                     get_stock_loan_payments,       \
-                     get_subscription_fees,         \
-                     get_referrals,                 \
-                     get_day_trades,                \
-                     get_wire_transfers,            \
-                     get_margin_calls,              \
-                     get_margin_interest,           \
-                     get_documents,                 \
-                     download_document,             \
-                     download_all_documents,        \
-                     get_all_watchlists,            \
-                     get_watchlist_by_name,         \
-                     post_symbols_to_watchlist,     \
-                     delete_symbols_from_watchlist, \
-                     build_holdings,                \
-                     build_user_profile,            \
-                     load_phoenix_account,          \
-                     withdrawl_funds_to_bank_account, \
-                     deposit_funds_to_robinhood_account
-
-from .authentication import login, \
-                            logout
-
-from .crypto import load_crypto_profile,        \
-                    get_crypto_currency_pairs,  \
-                    get_crypto_info,            \
-                    get_crypto_quote,           \
-                    get_crypto_quote_from_id,   \
-                    get_crypto_positions,       \
-                    get_crypto_historicals
-
-from .export import export_completed_stock_orders, \
-                    export_completed_option_orders,\
-                    export_completed_crypto_orders
-
-from .helper import request_get,      \
-                    request_post,     \
-                    request_delete,   \
-                    request_document, \
-                    update_session,   \
-                    set_output,       \
-                    get_output,       \
-                    filter_data                       
-
-from .markets import get_currency_pairs,                    \
-                     get_markets,                           \
-                     get_market_today_hours,                \
-                     get_market_next_open_hours,            \
-                     get_market_next_open_hours_after_date, \
-                     get_market_hours,                      \
-                     get_all_stocks_from_market_tag,        \
-                     get_top_movers,                        \
-                     get_top_movers_sp500,                  \
-                     get_top_100
-
-from .options import get_aggregate_positions,                           \
-                     get_market_options,                                \
-                     get_all_option_positions,                          \
-                     get_open_option_positions,                         \
-                     get_chains,                                        \
-                     find_tradable_options,                             \
-                     find_options_by_expiration,                        \
-                     find_options_by_strike,                            \
-                     find_options_by_expiration_and_strike,             \
-                     find_options_by_specific_profitability,            \
-                     get_option_market_data_by_id,                      \
-                     get_option_market_data,                            \
-                     get_option_instrument_data_by_id,                  \
-                     get_option_instrument_data,                        \
-                     get_option_historicals
-
-from .orders import get_all_stock_orders,               \
-                    get_all_option_orders,              \
-                    get_all_crypto_orders,              \
-                    get_all_open_stock_orders,          \
-                    get_all_open_option_orders,         \
-                    get_all_open_crypto_orders,         \
-                    get_stock_order_info,               \
-                    get_option_order_info,              \
-                    get_crypto_order_info,              \
-                    find_stock_orders,                  \
-                    cancel_all_stock_orders,            \
-                    cancel_all_option_orders,           \
-                    cancel_all_crypto_orders,           \
-                    cancel_stock_order,                 \
-                    cancel_option_order,                \
-                    cancel_crypto_order,                \
-                    order,                              \
-                    order_buy_market,                   \
-                    order_buy_fractional_by_quantity,   \
-                    order_buy_fractional_by_price,      \
-                    order_buy_limit,                    \
-                    order_buy_stop_loss,                \
-                    order_buy_stop_limit,               \
-                    order_buy_trailing_stop,            \
-                    order_sell_market,                  \
-                    order_sell_fractional_by_quantity,  \
-                    order_sell_fractional_by_price,     \
-                    order_sell_limit,                   \
-                    order_sell_stop_loss,               \
-                    order_sell_stop_limit,              \
-                    order_sell_trailing_stop,           \
-                    order_buy_option_stop_limit,        \
-                    order_sell_option_stop_limit,       \
-                    order_buy_option_limit,             \
-                    order_sell_option_limit,            \
-                    order_option_spread,                \
-                    order_option_credit_spread,         \
-                    order_option_debit_spread,          \
-                    order_buy_crypto_by_price,          \
-                    order_buy_crypto_by_quantity,       \
-                    order_buy_crypto_limit,             \
-                    order_buy_crypto_limit_by_price,    \
-                    order_sell_crypto_by_price,         \
-                    order_sell_crypto_by_quantity,      \
-                    order_sell_crypto_limit,            \
-                    order_sell_crypto_limit_by_price,   \
-                    order_crypto
-
-from .profiles import load_account_profile,     \
-                      load_basic_profile,       \
-                      load_investment_profile,  \
-                      load_portfolio_profile,   \
-                      load_security_profile,    \
-                      load_user_profile
-
-from .stocks import get_quotes,                 \
-                    get_fundamentals,           \
-                    get_instruments_by_symbols, \
-                    get_instrument_by_url,      \
-                    get_latest_price,           \
-                    get_name_by_symbol,         \
-                    get_name_by_url,            \
-                    get_symbol_by_url,          \
-                    get_ratings,                \
-                    get_events,                 \
-                    get_earnings,               \
-                    get_news,                   \
-                    get_splits,                 \
-                    find_instrument_data,       \
-                    get_stock_historicals,      \
-                    get_pricebook_by_id,        \
-                    get_pricebook_by_symbol,    \
-                    get_stock_quote_by_id,      \
-                    get_stock_quote_by_symbol
