+++ robin_stocks/robinhood/crypto.py
@@ -1,6 +1,6 @@
 """Contains functions to get information about crypto-currencies."""
-import robin_stocks.helper as helper
-import robin_stocks.urls as urls
+import robin_stocks.robinhood.helper as helper
+import robin_stocks.robinhood.urls as urls
 
 @helper.login_required
 def load_crypto_profile(info=None):
