+++ robin_stocks/robinhood/helper.py
@@ -10,7 +10,7 @@ Functions
 from functools import wraps
 
 import requests
-from robin_stocks.globals import LOGGED_IN, SESSION, OUTPUT
+from robin_stocks.robinhood.globals import LOGGED_IN, SESSION, OUTPUT
 
 
 def set_login_state(logged_in):
